Reorganize collection entity file
