+++ frontend/src/metabase/entities/collections/index.js
@@ -1,2 +1,5 @@
+export * from "./constants";
 export * from "./collections";
+export * from "./utils";
+
 export { default } from "./collections";
